nativeclone: support non-eden working copies
Summary:
Hook the Rust clone code in to the working copy init. I tweaked the arg checking to not require "--noupdate".

As you can see in test-hgrc.t, we no longer use a user friendly template when writing the repo's .hg/hgrc. Also visible in test-hgrc.t, I this fixes a regression in the handling of "%" in repo URLs.

Reviewed By: quark-zju

Differential Revision: D35985475

fbshipit-source-id: d10a9589fb941d9fc0888ad0aba1f4e11ba47794