nativeclone: infer repo name from source
Summary:
Now we infer the repo name from the specified repo source URL using the existing configparser logic. We use this repo name to load the repo specific config file, if present, among other things. This obviates the need to specify --configfile when cloning in many cases.

I had to move the edenapi capabilities check after the repo name is inferred since that depends on the repo name being set.

Reviewed By: DurhamG

Differential Revision: D36070460

fbshipit-source-id: 2e2a03933161713851a38979d3dd8b61ccba27df