allow redirections on read only mounts
Summary:
When eden mounts are marked readonly we are not able to remove the directories
underneath the bind mounts, but we are able to mount a bindmount over a
directory.

Not having the redirects means that buck can not be used on mounts in readonly
mode.

Removing the directory under a linux bind mount does not effect the bind mount,
so let's skip this step so long as the directory is only an empty directory.

Reviewed By: xavierd

Differential Revision: D36363079

fbshipit-source-id: 90c292bbe428a1e8d332d3cd81a93841b7a5f217