hgcommands: provide a synopsis function for Rust commands
Summary:
Makes Rust commands provide a `synopsis` function, which should optionally return a string for providing help about the positional arguments of the command.

For most commands a dummy function has been provided, but for the `clone` and `status` commands, the help for the positional arguments has been taken from their Python equivalents.

Differential Revision: D36362080

fbshipit-source-id: 8181c8b831f09111160de6a88fa32517a34fe809