repo: clean up repo init a bit
Summary:
It was uncomfortable how Repo::init took in an arbitrary mutable ConfigSet. It would do a config.load() to re-load the repo's config, thus mixing the two together. Depending on what was in the config passed in you could get a Repo back with a different config than what a subsequent Repo::load would give you.

Tweak things so that Repo::init does not take a mutable config. It writes out the various files and then does a normal Repo::load(). For now it is the caller's responsibility to mix any --config values back in.

Note that init_hg_repo no longer triggers any config loading. Previously it loaded the hgrc to trigger a dynamic config load so that, for example, logic in write_store_requirements had a full dynamic config available. This was somewhat problematic because reloading the config will clobber "--config" CLI flags that conflict with values in the dynamic config (or in other config files being loaded). The config in-hand does already contain the non-repo specific dynamic config, but if we want repo-specific values to be available we will need to either re-load the config first in the clone command once we have figured out the repo name, or build better Rust support for "read only" config items so we know how to preserve --config and --configfile values across reloads.

I also made Repo::init take ownership of (and drop) the passed in config. This is to encourage the caller of Repo::init to use the returned repo's config instead of the passed config.

Reviewed By: sggutier

Differential Revision: D36324334

fbshipit-source-id: 8737f9bc89a6034c5063b5c3a223e917e61784cf