sparse: separate root and non-root profile
Summary:
Add a new type sparse::Root which wraps sparse::Profile. Root represents the root profile (normally contained in .hg/sparse). Users of this library should only ever deal with the root profile, so this clarifies the API.

The root profile has special semantics for evaluating child profiles when creating the sparse matcher, so it is distinct from other profiles.

Reviewed By: DurhamG

Differential Revision: D36289459

fbshipit-source-id: e434ca92f286657c559a11baa9db73a1f05ba343