setdiscovery: include heads in the master group as "locally known"
Summary:
With lazy changelog, the pull code path first uses an optional fast path that
adds lazy segments to the master group, then uses the regular pull.

The regular pull by default will use visible heads + remote names to decide
what commits are present in the local repo. The lazy segments just pulled via
the fast path are not included in the setdiscovery step. That forces the server
to provide commits that were already pulled via the fast path.

Fix it by considering the heads of the master group that are not yet referred
by remotenames or visibleheads as locally known.

Reviewed By: mitrandir77

Differential Revision: D36335928

fbshipit-source-id: 54ea0f78a86c87e9ea4eb165cd862feb509f55cd