help: fix wrong documentation when running hg help
Summary:
When one tries to get the documentation for commands that exist both in Rust and Python (such as `status` and `clone`), the help documentation for the python one is shown instead of the Rust one. If we have both Rust and Python commands, we should default to the Rust ones.

This also gets rid of the duplicated messages.

Differential Revision: D36173444

fbshipit-source-id: d1bfc8985a0c1edfc5fa9034e1693f811b36c73a