Count time on critical pushrebase path
Summary:
This adds an ODS counter that adds up all time used on the critical land path for pushrebase. I based myself on some other ODS logging I've seen around mononoke.

We don't lock and instead do optimistic writes and retry on failure. The critical path is the time from when we read the bookmark for the first time, until the time we change the bookmark, and only if that bookmark move succeeded.

Reviewed By: mitrandir77

Differential Revision: D36317478

fbshipit-source-id: c5ff0c40ee6c73b73f2c27c3186a83ec67096e82