service: fix user after free in background glob
Summary:
When glob is run in the background, we need to make sure that the
ObjectFetchContext is kept alive for the whole duration of the glob.
Unfortunately, the code only kept it alive for the duration of the Future
returned by the globber, which for a background glob is a Future returning an
empty glob, not the background glob.

To fix this, we can simply move the background logic back into the caller where
the ObjectFetchContext can be kept alive with a simple `.ensure`.

Reviewed By: chadaustin, genevievehelsel

Differential Revision: D36482903

fbshipit-source-id: ee01fd664ea8933ada7f943128eab9eb5b754a27