try linking libatomic at the end of the link command rather than the beginning
Summary:
The reason we can't use -latomic as a linker flag is because it must show up on the link line after libraries that depend on it. Instead, have every executable and shared library in the EdenFS CMake build pull in libatomic.

This fixes the EdenFS CMake build on some systems.

Reviewed By: xavierd

Differential Revision: D36450490

fbshipit-source-id: 2fe7dc4497554dc8e009cb3c744695684ff08f6e