Use mapping prefix in DMv2
Summary:
Needs to be landed after D36445248.

See the comment on the code I'm deleting. We were ignoring the prefix so that the first derivation didn't have any prefix. Since now we've moved everything to DMv2, we can delete the current one from config, and start considering it again in code.

Reviewed By: farnz

Differential Revision: D36445291

fbshipit-source-id: 6615e5783f09fb9351c8459bec6aa68fda425aa7