Remove DeletedManifestVersion
Summary:
This removes all deleted manifest version code. After it is deployed everywhere, we can remove the field from configerator.

DMv1 code still exists but needs to be explicitly used. It will be deleted on the following diff.

The other option here would be to keep the enum, but only the V2 variant, which in my opinion just makes things ugly, so let's remove it alltogether.

This assumes everything already moved to DMv2, which is true, but to be safe I'll wait a while before landing this.

Reviewed By: kris1319

Differential Revision: D36444986

fbshipit-source-id: 07b149dfb32f535347f7345af1943fcfec7cf2b9