store: make the ObjectStore aware of case sensitivity
Summary:
One of the not well documented and already broken assumption in EdenFS is that
the ordering of TreeEntry in a Tree and the ordering of DirEntry in a TreeInode
is the same. Both the diff code and the checkout code relies on it to be the
same to decide whether entries are the same, or added, or removed.

The issue is that the ordering in a Tree is first not guaranteed, second never
checked and third fully dependant on the BackingStore which may or may not
respect this property. Today, we're somewhat lucky that the ordering is more or
less the same. It however breaks down on case insensitive mounts, where
insertion of an entry into a TreeInode heavily depends on the order and the
case!

Fixing this is unfortunately not as simple as just the insertion order, as this
breaks the ordering assumption of TreeEntry, and thus both checkout and diff
fail spectacularly. The proper fix is to ensure that TreeEntry are stored in a
PathMap, and then the PathMap can be fixed. A first step towards this goal is
to make sure Tree are case aware and thus needs to be constructed with a
CaseSensitivity argument. For now, this is only used in getEntryPtr, but in the
future, it'll be used to construct the PathMap.

One downside of adding an argument to the Tree constructor is that all call
sites needs updating. For now, I've made the choice of not doing this, and
instead hacking it up in the entry point for the backingstore: the ObjectStore.
This is extremely ugly, but is very pragmatic: on Windows and macOS, conversion
of case sensitivity will almost never happen.

In subsequent diffs, the BackingStores, LocalStore and TreeCache will be
converted to build Tree with the right case from the beginning.

Reviewed By: chadaustin

Differential Revision: D36257251

fbshipit-source-id: 17590f9a9f3608fa3a3dbef39b10eefbece4c745