model: use PathMap in Tree
Summary:
One of the unchecked and not documentation assumption of Tree is that they must
be sorted the same way as a TreeInode. Unfortunately, there is no guarantee
that the TreeEntry in a Tree are sorted, and the only place where this
guarantee is upheld... is in the testharness test library.

To solve this, the best way is to use the same underlying datastructure in both
TreeInode and Tree: a PathMap. This is guarantee to be sorted, and it properly
takes into account case sensistivity.

Reviewed By: chadaustin

Differential Revision: D36257252

fbshipit-source-id: 0fadd3c84d8e95631a8293c6797544ea94046a10