Make it possible to "reblame" efficiently when doing mutable blame
Summary:
When an ancestor has new blame due to a mutable rename, we don't want to have to re-diff the entire stack. Make it possible to fix blame based on two blame blobs (old and new) for the same location in history.

Core concept is that we can identify the parts of blame in the current place that haven't changed since the old blob was generated, and rewrite it to look like blame was always the "new" blob.

Reviewed By: markbt

Differential Revision: D36014998

fbshipit-source-id: ee46651f71ce3b84a6da13b620df1b421cea030d