Add logic to run_hooks endpoint
Summary:
This adds basic logic to the `commit_run_hooks` endpoint.

It takes a bookmark name, runs all the default hooks for landing the given diff at the bookmark, and returns the results as a list of outcomes.

This seems to do a good enough job as it is, but I will also add pushvars in the next diff.

Reviewed By: farnz

Differential Revision: D36515627

fbshipit-source-id: efcedaa57fb66ee24ef3472ae3cb9d9bb36eb166