amend: remove O(repo) children computation
Summary:
hg next/prev had logic that would iterate over every descendant of the
commit.  If you were on a commit early in the repository, this could be a very
large number of commits to iterate over.

With segmented changelog, computing children is O(merge) instead of
O(descendants), so let's just compute children the normal way now.

Reviewed By: quark-zju

Differential Revision: D36462672

fbshipit-source-id: 0549e27968d4b94e8a7efd4c4c8d16dd5f5d96fd