mononoke_api: add RepoConfig to InnerRepo
Summary:
`RepoConfig` is a facet, but currently not used that way.  Move it into
`InnerRepo` so we can access it via the `RepoConfigRef` trait.

The LFS tests currently mutate the repo config.  This is a bad pattern to use,
as parts of the repo config may be cloned off elsewhere, so it should be
considered immutable once the repo has been constructed.  For now, add a
special constructor for these LFS tests (like xrepo).  In the future we should
make `mononoke_api` able to construct test repos using any config override.

Reviewed By: kris1319

Differential Revision: D36444243

fbshipit-source-id: 72c1e4f2f7585ba0a8e79a94df399938b5160b39