Pushrebase client
Summary:
This creates a `PushrebaseClient` trait that abstracts away part of the pushrebase process.

In the future, this trait will be used to allow for remote pushrebasing through SCS. On this diff, however, it's merely a refactor, I created a `LocalPushrebaseClient` that simply does exactly the same that happened before.

Reviewed By: farnz

Differential Revision: D36598957

fbshipit-source-id: f7bcf6f2b0aec7428cdd6f21884658153430dd77