Mononoke Sharded Process Manager: Update tokio handle for spawning and blocking
Summary:
The `BackgroundProcessExecutor` framework relied on the current `tokio::Handle` to spawn tasks and to block futures for synchronous execution. However, the spawning and blocking happens on the `ShardManager` callback thread that crosses the C++ - Rust boundary and executes without a `tokio::Runtime` which leads to `panic` during spawning and blocking.
This diff includes the following changes:
- Updated the `BackgroundProcessExecutor` constructor to require a `tokio::Handle` from the caller.
- Updated `tokio::spawn` with `runtime_handle.spawn` and `Current.Handle.block_on` with `runtime_handle.block_on`.
- Updated all the call-sites to provide a reference to the underlying `tokio:Handle` while constructing the executor.

Reviewed By: markbt

Differential Revision: D36660061

fbshipit-source-id: 4ba94896ee1d47fd45a65282a872eeb8acd989d1