clone: add config for deprecating the --rev option
Summary:
The `--rev` option existed for pulling in some specific diff when doing a clone. Since this config has not been used by anyone in the last few months and it is not something we want to support from now on, we should start deprecating it. Before we outright remove it from the codebase, we are first going to be entirely sure that no one is using it, so we are going to roll out some config via dynamic config so that the clone command crashes if this flag is requested.

Additionally, the help text for this option was modified so that it wouldn't be shown when running `hg help clone`, unless the `--verbose` option was passed.

Reviewed By: DurhamG

Differential Revision: D36522176

fbshipit-source-id: 769a26f539124bc7049c01818b0d26442b1f4ade