pullcreatemarkers: force load marker targets into idmap
Summary:
The pullcreatemarkers code was not causing commits to show "Landed as".
This was because the "Landed as" logic depends on the obsolete computation
(https://www.internalfb.com/code/fbsource/[124e57ed2087addfbbbcef6ea38ed6f1a1569847]/fbcode/eden/scm/lib/mutationstore/src/lib.rs?lines=272)
and that computation ignores mutation markers if the marker successor hash is
not yet in the local idmap. This seems weird, but for now let's work around it.

This is combined with another issue where, even though the hash is added to the
idmap earlier in the logic, during the `unfi[publicnode]` lookup, it is thrown
away when we do unfi.lock() because that invalidates the current repo state.
Ideally we'd flush the idmap before we invalidate the state, but I don't believe
theres a way to flush just the idmap, and I'm not certain we can safely flush
the rest of the dag.

So for now, we force load all the public nodes again inside the lock.

Reviewed By: quark-zju

Differential Revision: D36721868

fbshipit-source-id: 4898078862535782287f300ac1118b420d5b1a0d