Chunk changesets fetches from MySQL to push pain into our code
Summary:
kris1319's oncall report calls out that we have problems with very expensive `SelectManyChangesets` queries. https://fburl.com/daiquery/hgizg6mo shows me that the mean number of rows selected is 7, which is also the p5 rows, but that the median is 168 rows, and the p95 a whopping 148,961,263 rows.

As a band-aid, chunk our big fetches into blocks of 1,000 rows, and permit two parallel fetches. This won't fix the pain, but will result in our large requests filling the local cache in stages, and hopefully push the pain away from XDB and into Mononoke, where the request that triggered a huge fetch will still be slow, but we'll be filling our caches as we go along.

Reviewed By: mitrandir77

Differential Revision: D36762597

fbshipit-source-id: 08494c400debc84bea4899fa07fa35047ef617dd