mononoke: Add new RepoLock facet
Summary:
Mononoke currently reads and writes to the `repo_lock` table of the
database configured in `write_lock_db_address`. However, we want to migrate to
using a `repo_lock` table inside of the metadata DB.

Add a new `RepoLock` trait that can be used to check whether a repo is locked.
This is a trait so that there can be multiple implementors, such as the
`MutableRepoLock` that allows for updating the DB and the
`AlwaysLockedRepoLock` that is used when `readonly=true` is in a repo's config.

The existing `RepoReadWriteFetcher` struct is confusing as it tries to have the
logic for both of the scenarios I just mentioned.

NOTE: There is a behaviour change with the new `MutableRepoLock` logic - if
there isn't an entry in the DB then the repo will default to being unlocked.

Reviewed By: mitrandir77

Differential Revision: D36129276

fbshipit-source-id: b080b56fa2a8f55a98ad7b9737053a285b3e1a67