Migrate object parsing from git2 to gitoxide family crates
Summary: This stops us using libgit2 to parse and handle git objects, and leaves it as just a repo accessor. A later commit will change to using the git CLI to access the repo, and remove libgit2 from eden/mononoke code - pushing the compatibility layer into scm/grepo, which will be left until we're ready to migrate to live git import.

Reviewed By: mitrandir77

Differential Revision: D36712909

fbshipit-source-id: b827d51679e7667a0ff0bf8e11e70e96f1266e21