doctor: fix missing changelog for users with experimental.use-rust-changelog enabled
Summary: For deciding which implementation of the changelog should be used, the `localrepo._openchangelog` function expects the passed repo to contain an inner Rust repo object for users with the config mentioned in the title. The `doctor` command uses that function, but passes a repo object without the inner Rust repo, making it crash for users with the aforementioned config enabled. This diff fixes that by adding an inner Rust repo object to `ChangelogRepo`.

Reviewed By: jordanwebster

Differential Revision: D36776180

fbshipit-source-id: 9b65944ed7bccc71c8134ab6acc3a17700991ad1