repolock: add legacy lock file compat
Summary:
Tweak the rust repo lock code to also take the legacy Python lock. This is to support external programs such as Watchman and VS Code that look directly at the existence of .hg/wlock to determing if an hg operation is in progress. The Rust locking scheme is not compatible with that use case since the lock files are never deleted.

If the legacy file is locked, we don't mess with it. This avoids interfering with active Python locks in devel.lockmode=python_and_rust.

Ideally external programs would not depend on anything in .hg, but they do, and there is not a great alternative in this case.

Reviewed By: DurhamG

Differential Revision: D35943143

fbshipit-source-id: 8cadf53d5c18570b1cceec9564dc88880251de12