clidispatch: try network doctor if config load fails
Summary:
If a repo has no dynamic config cache, running any command with a bad network connection nets you an error trying to fetch the dynamic config. This error was too early for the network doctor to get involved because there is no config object.

In this diff, tweak a few things so the network doctor can maybe run in this case:
- Tag the dynamic config fetch error as a NetworkError. This is what the network doctor looks for.
- If we fail loading the repo's config, try again loading a repoless config. This will probably work as long as this isn't your first hg command ever on your machine since repoless hg commands will cache the dynamic config in the user's home directory.

Reviewed By: DurhamG

Differential Revision: D36530502

fbshipit-source-id: 9a671025d268c8cbe2cac80ca45a82eb202edcae