commands: change error code for failed fallback when forcing Rust
Summary:
Changes the error code from 255 to 197 and makes it not print anything when there is a failure on fallback.

We have planned making the hg-clone script look at the error code so that when it runs `hg clone` for detecting if some of the selected options for clone have not been implemented yet. By doing this, we can make hg-clone fall back to its old behavior.

This should be safe enough since we are not expecting the users to have this config enabled, nor we expect rolling it out to anyone.

The selected error code was 197 since this is the first illegal POSIX error code.

Reviewed By: DurhamG

Differential Revision: D36638023

fbshipit-source-id: 2bbe01b4a326d5ff1ff502e8b4aa7e103aee8bcb