clone: change clone.force-rust for commands.force-rust
Summary: Currently it is possible to force the Rust version of the clone command via the `clone.force-rust` config. Given that in the near future we will start migrating more commands to Rust, it would be convenient to have a universal switch for this behavior instead of having a config for each of our commands.

Differential Revision: D36776611

fbshipit-source-id: 253541f05a32c8f8b02478e72e734ee762f14467