Stop reading bundle_replay_data from SQL table
Summary:
With the rest of the stack deleting unused unbundle usage, we're rid of all but 1 usage of `bundle_replay_data` SQL table, so let's stop reading from it.

The only usage that is left is in backsyncer. It is a piece of code that both reads the unbundle data from the table, and then rewrites it to the table again. So removing it is fine and causes no issues.

We're also now avoiding doing a `JOIN` on a SQL query, so it should be more efficient :)

Reviewed By: mitrandir77

Differential Revision: D36704713

fbshipit-source-id: 1f09a1d8faab1abff1f2b159fc616cf3a1293f60