mononoke: Move disabled_hooks to MononokeEnvironment
Summary:
In order to build a `HookManager`, a `HashSet` of hooks to be disabled
is required. This is currently provided as part of the `MononokeApiEnvironment`
that some binaries populate. However, `MononokeApiEnvironment` contains a
`RepoFactory`, which means it is created after the `RepoFactory`. As a result,
the facet builder code doesn't have the disabled hooks information available.

Move `disabled_hooks` to the `Environment` struct and add a new `AppExtension`
that mutates the `Environment` if hooks are disabled on the command line.

Reviewed By: markbt

Differential Revision: D36809932

fbshipit-source-id: 0d2a31b4eec9a608b0fba5c3d98b669771a3c79c