mononoke: Convert test fixtures to not require BlobRepo
Summary:
In a later diff I am going to migrate the hooks tests to stop using
`BlobRepo` and instead use a container of facets. In order to do that, the test
fixtures relied on by the tests need updating to no longer require a
`BlobRepo`.

This diff updates fixture methods (such as `set_bookmark`) and the
`TestRepoFixture` trait to accept `&impl Repo`. The diff also adds a new
`TestRepo` struct that implements `Repo`, as a concrete type is required as a return type from a trait method.

Finally, `TestRepoFixture` is updated to include a `get_test_repo` method that
returns `TestRepo`. This will allow individual tests to opt out of using
`BlobRepo`.

NOTE: In the future `TestRepo` will probably be removed and groups of tests
will be able to define their own facet containers. However, `TestRepo` has been
created to make the migration away from `BlobRepo` easier.

NOTE: I've had to update a few tests to include types to keep the compiler
happy. I decided to use `BlobRepo` in these cases to keep this change small.

Reviewed By: markbt

Differential Revision: D36939590

fbshipit-source-id: acde236ce853e3e7bc0755baa3fbd05ec0682a9f