Disable SQL lag monitoring, with tunable to re-enable
Summary:
Talking to SQL team, we found out that the `replication_status` call fails very often, and causes a lot of alarms for them. Our SQL tables are simply too big for it, and we need to remove this or figure out a better way to do it.

This call is used in two places:
- Delaying requests to blobstore. The failures here don't affect us because we simply [ignore the errors](https://fburl.com/code/buf6o9c4) (without logging).
- Waiting for replication lag to go down before continuing some operation. This is used in blobstore healer, segmented changelog tailer, and megarepo commit rewriting. This part can actually fail if the replication_status request fails.

The approach decided was to simply remove this logic and let SQL handle it. If it becomes a problem and we actually need some throttling ourselves, SQL team will help us come up with something that can work with our size.

I gated this behind a tunable (which disables lag monitoring by default). So if it brings problems, we just need a quick configerator diff to mitigate it. I'll post on SC rollouts before landing.

Reviewed By: farnz

Differential Revision: D37001468

fbshipit-source-id: ed92567284f0d730715f0a4cf89b58c9493a7afb