default to loading blobs and trees from the BackingStore in debug blob and debug tree commands
Summary:
I spent some time debugging the contents of a blob, and the fact that
we don't load blobs from the BackingStore was confusing, especially
since the hg backing store no longer uses blob caching. Change the
default behavior to load from the BackingStore, but add the ability to
opt out with --no-load.

Reviewed By: genevievehelsel

Differential Revision: D36949521

fbshipit-source-id: 2a7f5abcb70513699b410bcadd49113b5c346ff4