assert implicit InodeOrTreeOrEntry invariant
Summary:
We don't contruct InodeOrTreeOrEntry from a TreeEntry that represents a tree.
Instead we resolve it to a tree before we create an InodeOrTreeOrEntry for it.

We have a couple place where we comment about this, and we rely on this
assumption in getorFindChild.

However, it is not asserted in the code. Lets make an assertion, so its
less likely we accidentally break this invariant.

Reviewed By: xavierd

Differential Revision: D37007606

fbshipit-source-id: 2916f0d3348e2b3df6b9042e24a1045cfa164a78