Edited eden trace inode subcommand to call C++ trace_stream binary
Summary: Since at this point there are no tools to retrospect on past EdenFS events if debugging tools were not set up previously, we wish to add --retroactive options to EdenFS tracing commands to help debug such events that occurred even if debugging tools were not set up at the time. We first wish to set this up with the trace inode subcommand and here we continue building out scaffolding for this subcommand (building off of D36910336 (https://github.com/facebookexperimental/eden/commit/10b9fe6e5955c8922edc21287d012a8021ee47e8)). Specifically we edit the command to call into C++ code in the trace_stream binary. This is done so that the code for the trace inode command remains consistent and has the same shape as the other eden tracing commands.

Reviewed By: genevievehelsel

Differential Revision: D36939257

fbshipit-source-id: 7224873f687a292b4f9b9458bc3dc6d5b83d3756