status: naive implementation of watchman merge to return pending changes
Summary:
A very naive implementation of merge that gets the tests passing.

Later we can optimize this and start leveraging additional information that watchman can provide to us.

I've modified the API so that merge is responsible for writing to tree state and returning the pending_changes, making this a single-use struct. This more accurately reflects the intended usage of this and makes state management much simpler than having three methods: merge/persist/pending_changes.

Reviewed By: DurhamG

Differential Revision: D36431980

fbshipit-source-id: de78db3d3807f0de625418364f3893a3c76599e9