Stop writing bundle_replay_data to SQL table
Summary:
This diff makes mononoke stop writing bundle_replay_data. All uses of it were removed on the previous diff (and they already aren't used anywhere in prod anyway).

It updates the Bookmark traits so a bundle_replay_data doesn't need to be specified, which is a bit annoying because it's used A LOT, that's why I had to touch so many files to stop using this field.

To stop the diff from becoming even larger, I stopped at some places by allowing unused variables and adding a TODO, which I'll tackle on the rest of the stack.

The most important change is on the file that actually does the writing to DB, `eden/mononoke/bookmarks/dbbookmarks/src/transaction.rs`

Reviewed By: farnz

Differential Revision: D36707089

fbshipit-source-id: 81a03278206e28f902236018bed7e2fa3f2bf337