Added Test Suite for ActivityBuffer
Summary:
This change creates unit tests for the ActivityBuffer class. ActivityBuffer is a fixed size buffer that will store recent events that have happened within EdenFS (currently this consists of only InodeMaterializeEvents though we intend for this to be more general later), and the purpose of this class is to allow for easier debugging retroactively of past Eden events by reading from the ActivityBuffer. This change specifically adds three tests:
1. To test if an ActivityBuffer can be instantiated without errors
2. To test if adding to an ActivityBuffer indeed adds it to the buffer's list of events
3. To test if events get removed when the buffer reaches its size limit

Reviewed By: MichaelCuevas, kmancini

Differential Revision: D37008866

fbshipit-source-id: 3316898ede317bd648a13e7d5cc308d9267c4209