Create Configuration Option for the Max Number of Events in an ActivityBuffer
Summary:
In order to more easily retrospect on past events occurring in EdenFS, we are building out a new ActivityBuffer that will be used in retroactive versions of our tracing CLI. The ActivityBuffer will enable access to a fixed size amount of recently stored events that have happened within EdenFS for users to view and be able to debug EdenFS issues more easily with.

In a previous diff (D36904356 (https://github.com/facebookexperimental/eden/commit/408dd659fbee4ccab5df60026fe7f6f59f01c708)), we added a configuration setting to enable/disable use of the ActivityBuffer, and in this diff I am introducing a new configuration option for the maximum number of events that can be stored by the ActivityBuffer before old events begin to get evicted.

Reviewed By: kmancini

Differential Revision: D37043817

fbshipit-source-id: 2c5cb36cd7277500ab019b58a0405819b99093f3