status: Persist latest state to treestate after watchman update
Summary:
Set clock, mark files as need check and clear needs check on files that are no longer modified.

Realised a pretty severe flaw in filechangedetector: we need to also return the unchanged files in resolve_maybes so that we can clean up their treestate. Will fix that in a later diff.

Reviewed By: DurhamG

Differential Revision: D37040605

fbshipit-source-id: 26b1ce15eb884d210465024a7337a2dd2d44f07f