context: allow memctx to parent on memctx's
Summary:
Currently the memctx constructor takes parent nodes, which means a
memctx can't be on top of another memctx. Let's change it to take a context
instead.

This is causing a bug in 'hg amend --to' which expects to be able to create a
stack of memctxs and perform a rebase like operation onto them.

Reviewed By: quark-zju

Differential Revision: D36872499

fbshipit-source-id: 04a3d0ab6841f5e125592da033013591e890ef72