amend: fix --to on a stack
Summary:
There was a bug where amending to file X would undo the change later in
the stack if another commit touched file X. This was because each memctx in the
stack we created was parented on it's original commit, and not the previous
commit in the memctx stack. This caused it to read the original file contents
for any file not modified in that particular memctx.

Since memctx's can now be parented on other memctxs (as of the last diff), let's
change amend --to to use that.

Differential Revision: D36872498

fbshipit-source-id: fe2ff322a8d836c892a56ec9313315b0f43336d0