Implement the ActivityBuffer
Summary:
ActivityBuffer is a class that will store events of past changes within EdenFS in a fix sized buffer (currently these events only consist of InodeMaterializeEvents but we intend to generalize this to other events later). This class will be used in order to enable retroactive debugging of EdenFS events through new retroactive versions of the Eden tracing CLI that will access this buffer, allowing for easier debugging of difficult or expensive to reproduce issues.

This change implements the ActivityBuffer class, specifically implementing the constructor, addEvent, and getAllEvents methods that initialize the class, add an event to it (while evicting an event if the buffer reaches its max size), and returns a std::list of all stored events respectively in a thread safe way using folly::Synchronized.

Reviewed By: MichaelCuevas

Differential Revision: D37040995

fbshipit-source-id: 98869a5c1a46f0d62ed31027abdda7b264775515