Handle sparse profile config change
Summary: Changes in sparse profiles configs themselves are tricky. We don't know how particular change affects the profile, so let's recalculate the size from scratch. Another thing is that profiles can be moved or copied. We have designated location `fbcode/tools/scm/sparse` for profiles config. (This will be configured in the future diffs) So the only profiles in that directory will be monitored. If profile is moved out from that directory we will consider it as removed profile.

Reviewed By: yancouto

Differential Revision: D36695690

fbshipit-source-id: 8d7646c8e7a3a0f1b5b90359cd823cc57c111519