pushrebase_client: allow overriding of scs client certs
Summary:
Allow the Mononoke pushrebase client to override which certificates it
will use when connecting to the Source Control Service.

There is, unfortunately, no convenient place to store these paths.  Connections
to SCS are made from within free functions within `repo_client`, so we don't
have a global context object to store them on (and threading them through the
call stack is a massive amount of work).   To be pragmatic, the override
applies to all SCS clients within the same process.  This is the same behaviour
as the `THRIFT_TLS_CL_` environment variables, but scoped specifically to the SCS
client connections.

Reviewed By: yancouto

Differential Revision: D37044084

fbshipit-source-id: 262c79ba544f7199bb7fbbaefd682bb68ffe32dc