tests: add a script to batch editing test headers
Summary:
There are some common compatibility headers tests use. Sometimes they need to
be edited in batch - for example, take the test runner result and append a
feature header to failed tests and try again. There is no convenient way to do
that. `features.py` is a way but is easy to forget for new tests.

Add a script to edit the feature headers. Hopefully it can replace the use of
`features.py`.

Reviewed By: DurhamG

Differential Revision: D34931994

fbshipit-source-id: d1f769db5a66393819624541f9eeb2f4d90f9314