match: fix empty listfile to not match everything
Summary: Tweak things so specifying an empty listfile doesn't end up matching everything. I don't think the code in question here was expecting the case where you start with a non-zero amount of patterns but end up with an empty list of patterns. It is a pretty niche case (AFAIK only an empty listfile would cause it), so I'm hoping it won't break anything else.

Reviewed By: DurhamG

Differential Revision: D37051913

fbshipit-source-id: 19537f2e14da7646c027c4e3fc90f3d92abf5f06