wal: implement reads and add tests
Summary:
This diff introduces WAL read-query, that will be later used by the healer job to fetch new entries from the blobstore queue.

The main difference between the sync-queue and WAL structure is that there is only one entry per `put`-request in case of WAL, while for the sync-queue there was an entry per each underlying blobstore, so the re were multiple rows per `operation_key` in the table.
Because of that and the fact that healer by design operates on batches of `put` operations, the read-query was very expensive: it required to filter out particular number of distinct `operation_keys` and then select the rows that correspond to those keys.

This implementation should be much more efficient: just getting `limit` number of rows from Mysql table will give the healer `limit` number of distinct `put` operations to work on, by design.

Reviewed By: mitrandir77

Differential Revision: D37099369

fbshipit-source-id: b82a3682b91967c1cffd20b6560898bbba247c5d