wal: implement writes
Summary:
We want the writes to the queue to happen as soon as possible, so we're going to enqueue new entries while there is a write-query to Mysql in-flight. Then the WAL will fetch batch of the enqueued items and write those to Mysql.

The basic implementation idea is to use mpsc channel so the clients could write new entries to the channel, while the WAL will read and process them. On the completion, when the entry is finally written to the Mysql table, the clients will be notified of the success or error over the oneshot channel.

Generally this is the same approach the current blobstore sync-queue uses.

Reviewed By: mitrandir77

Differential Revision: D37069878

fbshipit-source-id: c8313e99f2bd7b830c8ed6ce9ff739c31492c333