TreeInode expose readdir for thrift
Summary:
Buck2 needs to be able to list directory contents to search through the
filesystem for targets. It can use the filesystem for this, but on NFS
readdir can be suprisingly expensive, as many standard libraies implementation
of readdir will endup causing EdenFS to load all the children in the directory.

Here we add a method to Tree inode to get a list of its children.

Note that we keep this method seperate from the existing readdirImpl
implementation that the filesystem channels use because for now it does
not need to be incremental, and we don't need information about our parent inode.

Reviewed By: chadaustin

Differential Revision: D36986975

fbshipit-source-id: 7306c9ef602ef975837b4f20ab1b584e1128fb0d